Move beacon-chain-handler/lib.rs into producer.rs submodule in preparation for importer.rs submodule
