+++ node/beacon-chain-handler/Cargo.toml
@@ -8,5 +8,6 @@ chain = { path = "../../core/chain" }
 primitives = { path = "../../core/primitives" }
 node-runtime = { path = "../runtime" }
 storage = { path = "../../core/storage" }
-
 parking_lot = "0.6"
+futures = "0.1.25"
+tokio = "0.1.11"
